refactor : 기능 분리 , 제약 조건 추가
